requirements.txt
django==4.2.4
psycopg2-binary==2.9.7
pillow==10.0.0
djangorestframework==3.14.0
python-dotenv==1.0.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
stripe==6.5.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.0